update for compatibility with rewrite 7.35.0
